refactor(plugin-p6): 风控中心前端搬家——View/spec/api 纯移动迁入 plugins/content-moderation
Phase-6 TASK-004 前端半场：git mv 三个文件进 src/plugins/content-moderation/
（RiskControlView.vue / __tests__/RiskControlView.spec.ts / api.ts），零逻辑改动：

- 路由懒加载路径与 api/admin/index.ts 转发出口同步新位置
  （riskControlAPI 与类型 re-export 保留，其余消费方零改动）
- 路由 path=/admin/risk-control、菜单位置、i18n、插件门控全部不变

验证：vue-tsc 无错；vitest 33 文件 218 用例全绿（含随迁 spec/pluginkit
守护/路由/侧边栏/i18n locale 套件）
