+++ frontend/src/api/admin/index.ts
@@ -31,7 +31,7 @@ import channelMonitorAPI from './channelMonitor'
 import channelMonitorTemplateAPI from './channelMonitorTemplate'
 import adminPaymentAPI from './payment'
 import affiliatesAPI from './affiliates'
-import riskControlAPI from './riskControl'
+import riskControlAPI from '@/plugins/content-moderation/api'
 import adminComplianceAPI from './compliance'
 
 /**
@@ -110,4 +110,4 @@ export type { BalanceHistoryItem } from './users'
 export type { ErrorPassthroughRule, CreateRuleRequest, UpdateRuleRequest } from './errorPassthrough'
 export type { BackupAgentHealth, DataManagementConfig } from './dataManagement'
 export type { TLSFingerprintProfile, CreateProfileRequest, UpdateProfileRequest } from './tlsFingerprintProfile'
-export type { ContentModerationConfig, ContentModerationLog, ModerationMode } from './riskControl'
+export type { ContentModerationConfig, ContentModerationLog, ModerationMode } from '@/plugins/content-moderation/api'
