+++ frontend/src/plugins/content-moderation/__tests__/RiskControlView.spec.ts
@@ -4,7 +4,7 @@ import { flushPromises, mount } from '@vue/test-utils'
 import type { DOMWrapper, VueWrapper } from '@vue/test-utils'
 
 import RiskControlView from '../RiskControlView.vue'
-import type { ContentModerationConfig, UpdateContentModerationConfig } from '@/api/admin/riskControl'
+import type { ContentModerationConfig, UpdateContentModerationConfig } from '@/plugins/content-moderation/api'
 
 const {
   getConfig,
